+++ src/transformer/tests/metadata.test.ts
@@ -5,9 +5,13 @@ import ts from 'typescript';
 import {
     F_REST, F_OPTIONAL, F_PRIVATE, F_PROTECTED, F_PUBLIC, F_READONLY, F_INFERRED, F_ABSTRACT, F_ARROW_FUNCTION,
     F_ASYNC, F_CLASS, F_EXPORTED, F_FUNCTION, F_INTERFACE, F_METHOD, F_STATIC, T_ANY, T_ARRAY, T_FALSE,
+    F_ARRAY_BINDING, F_OBJECT_BINDING,
+
     T_GENERIC, T_INTERSECTION, T_MAPPED, T_NULL, T_THIS, T_TRUE, T_TUPLE, T_UNDEFINED, T_UNION, T_UNKNOWN,
-    T_OBJECT, T_VOID, T_ENUM, T_FUNCTION, InterfaceToken, RtObjectMember, RtObjectType, RtFunctionType,
-    RtUnionType
+    T_OBJECT, T_VOID, T_ENUM, T_FUNCTION,
+
+    InterfaceToken, RtObjectMember, RtObjectType, RtFunctionType,
+    RtParameter, RtArrayType
 } from '../../common/format';
 import { runSimple } from '../../runner.test';
 
@@ -866,6 +870,50 @@ describe('rt:p', it => {
 
         let params = Reflect.getMetadata('rt:p', exports.foo);
         expect(params[0].v()).to.be.instanceOf(exports.A);
+    });
+    it('emits for array destructuring', async () => {
+        let exports = await runSimple({
+            code: `
+                export function foo(foo, [bar, baz]: string[]) { }
+            `
+        });
+
+        let params: RtParameter[] = Reflect.getMetadata('rt:p', exports.foo);
+        expect(params[0].n).to.equal('foo');
+        expect(params[0].f ?? '').not.to.contain(F_ARRAY_BINDING);
+        expect(params[0].f ?? '').not.to.contain(F_OBJECT_BINDING);
+        expect(params[1].f ?? '').to.contain(F_ARRAY_BINDING);
+        expect(params[1].f ?? '').not.to.contain(F_OBJECT_BINDING);
+        expect(params[1].n).not.to.exist;
+        expect(params[1].t()).to.eql(<RtArrayType>{ TΦ: T_ARRAY, e: String });
+        expect(Array.isArray(params[1].b)).to.be.true;
+        expect(params[1].b.length).to.equal(2);
+        expect(params[1].b[0].n).to.equal('bar');
+        expect(params[1].b[0].t()).to.equal(String);
+        expect(params[1].b[1].n).to.equal('baz');
+        expect(params[1].b[1].t()).to.equal(String);
+    });
+    it('emits for object destructuring', async () => {
+        let exports = await runSimple({
+            code: `
+                export function foo(foo, {bar, baz}: { bar: string, baz: number }) { }
+            `
+        });
+
+        let params: RtParameter[] = Reflect.getMetadata('rt:p', exports.foo);
+        expect(params[0].n).to.equal('foo');
+        expect(params[0].f ?? '').not.to.contain(F_ARRAY_BINDING);
+        expect(params[0].f ?? '').not.to.contain(F_OBJECT_BINDING);
+        expect(params[1].f ?? '').not.to.contain(F_ARRAY_BINDING);
+        expect(params[1].f ?? '').to.contain(F_OBJECT_BINDING);
+        expect(params[1].n).not.to.exist;
+        expect(params[1].t()).to.eql({ TΦ: "O", m: [{ n: "bar", f: "", t: String }, { n: "baz", f: "", t: Number }] });
+        expect(Array.isArray(params[1].b)).to.be.true;
+        expect(params[1].b.length).to.equal(2);
+        expect(params[1].b[0].n).to.equal('bar');
+        expect(params[1].b[0].t()).to.equal(String);
+        expect(params[1].b[1].n).to.equal('baz');
+        expect(params[1].b[1].t()).to.equal(Number);
     });
 });
 describe('rt:t', it => {